pyproject.toml
# See: https://github.com/Amourspirit/libreoffice-python-path-ext/wiki/pyproject.toml
[tool.oxt.config]
auto_install_in_site_packages = true
build_dir = "build"
dist_dir = "dist" # should be changed to someting like dist, tmp_dist is not included in git
install_wheel = false
oxt_name = "PyPath"
py_pkg_dir = "py_pkgs"
py_pkg_files = []
py_pkg_names = [] # ["ooodev", "ooo"]
token_file_ext = ["txt", "xml", "xcu", "xcs"]
update_file = "update.xml"
zip_preinstall_pure = true
window_timeout = 5 # number of seconds to wait for window to appear
dialog_desktop_owned = false # determines if the dialog is owned by the desktop window
resource_dir_name = "resources"
resource_properties_prefix = "pipstrings"
default_locale = ["en", "US"]

[tool.oxt.token]
# in the form of "token_name": "token_value"
# in the files the token will have a prefix and suffix of ___
pip_wheel_url = "https://files.pythonhosted.org/packages/50/c2/e06851e8cc28dcad7c155f4753da8833ac06a5c704c109313b8d5a62968a/pip-23.2.1-py3-none-any.whl"
test_internet_url = "https://duckduckgo.com"
log_pip_installs = false
lo_identifier = "com.github.amourspirit.extensions.pythonpath"
lo_implementation_name = "PyPathRunner"
display_name = "LibreOffice Python Path"
description = "Allows for including python packages in LibreOffice that are not included by default."
publisher = "AmourSpirit"
publisher_url = "https://github.com/Amourspirit/libreoffice-python-path-ext"
update_url_xml = "https://raw.githubusercontent.com/Amourspirit/libreoffice-python-path-ext/main/___dist_dir___/___update_file___"
update_url_oxt = "https://raw.githubusercontent.com/Amourspirit/libreoffice-python-path-ext/main/___dist_dir___/___oxt_name___.oxt"
url_pip = 'https://bootstrap.pypa.io/get-pip.py'
# set log_file to "" to disable logging
log_file = "pypath_install.log"
log_name = "LO PY Path"
log_level = "DEBUG" # NONE, DEBUG, INFO, WARNING, ERROR, CRITICAL
log_format = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
log_add_console = false # should be false for production
lo_pip = "lo_py_path"
platform = "all"
startup_event = "OnStartApp" # OnStartApp or onFirstVisibleTask
show_progress = false
delay_startup = false # determines if installing waits fo the window to load before installing

[tool.oxt.requirements]
# https://python-poetry.org/docs/dependency-specification/
# typing-extensions = ">=4.7.0"

[tool.oxt.preinstall.pure]
# verr = ">=1.1.2"

[tool.oxt.locale.publisher]
en = { name = "___publisher___", url="___publisher_url___" }

[tool.oxt.locale.name]
en = "___display_name___"
fr = "Chemin Python LibreOffice"
de = "LibreOffice Python Pfad"
pt = "Caminho Python LibreOffice"
es = "Ruta de Python de LibreOffice"

[tool.oxt.locale.desc]
en = "___description___"
fr = "Permet d'inclure des paquets python dans LibreOffice qui ne sont pas inclus par défaut."
de = "Ermöglicht das Einbinden von Python-Paketen in LibreOffice, die standardmäßig nicht enthalten sind."
pt = "Permite incluir pacotes python no LibreOffice que não estão incluídos por padrão."
es = "Permite incluir paquetes de python en LibreOffice que no están incluidos por defecto."


[tool.poetry]
name = "LO PY Path"
version = "1.2.3"
description = "Allows for including python packages in LibreOffice that are not included by default."
authors = [":Barry-Thomas-Paul: Moss <bigbytetech@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
ooouno = "^2.1.2"
ooo-dev-tools = "0.12.2"
toml = "^0.10.2"


[tool.poetry.group.dev.dependencies]
types-scriptforge = ">=2.0.1"
pytest = "^7.4.0"
pytest-dotenv = "^0.5.2"
black = "^23.7.0"
oooscript = "^1.1.3"
oooenv = "^0.2.2"
lxml = "^4.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylint.'MESSAGES CONTROL']
max-line-length = 120
disable = """
    no-name-in-module,
    no-self-argument
"""

[tool.pyright]
include = ["src"]
reportMissingModuleSource = false

[tool.ruff]
extended-select = ["ANN", "C4", "N", "SIM", "TCH"]
ignore = ["E501"]
show-fixes = true
target = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]

markers = [
    "skip_headless: skips a test in headless mode",
    "skip_not_headless_os: skips a test in GUI mode for give os",
]
